Data type

+++ src/BmiService.java
@@ -1,5 +1,7 @@
 public class BmiService {
-    public double calculate(double height, double weight) {
+    public int calculate(double height, double weight) {
+
+
         double bmi = weight / Math.pow(height, 2);
 
         return (int) bmi;
